Add an "experimental" API to send legacy telemetry
Unfortunately, we need to ship our own legacy telemetry APIs
in order for us to send schema name, namespace and other fields
which are expected to be set in the addon manifest. We did not
have this problem in about:ion, because we did not use this
https://searchfox.org/mozilla-central/rev/1a973762afcbc5066f73f1508b0c846872fe3952/toolkit/components/extensions/parent/ext-telemetry.js#64 API, but rather the `TelemetryController` directly.

The introduced API also provides a way to set and clear the Ion ID
in the user prefs, otherwise Legacy Telemetry will fail if that's missing.