Remove `DEFAULT_ARGS` from `core-addon/Core.js`
We had default values duplicated across that file
and the rollup configuration file, making it harder
to understand what was the real source of truth.
This makes the build system configuration file the
authoritative source.
Note that before this commit, the default values were
different in the two locations. This additionally drops
the `ION` prefix from the studies list name.