Fix tests and the config files (#334)
* Move `__ENABLE_DATA_SUBMISSION__` to the right config file

This was just being used in the rollupJS config used
for the Svelte part.

* Remove `DEFAULT_ARGS` from `core-addon/Core.js`

We had default values duplicated across that file
and the rollup configuration file, making it harder
to understand what was the real source of truth.
This makes the build system configuration file the
authoritative source.
Note that before this commit, the default values were
different in the two locations. This additionally drops
the `ION` prefix from the studies list name.

* Break tests on unhandled exceptions in MochaJS

This runs tests with ` --unhandled-rejections=strict` to
make sure that any uncaught exception doesn't go unnoticed.
Fixes #317.

* Use the correct studies-list option in the config

* Rename `__ION_WEBSITE_URL__` to `__WEBSITE_URL__`

* Fix the fallout from making mocha stricter

The test runner now breaks on unhandled
exceptions and turns out we have quite a few
of them. This commit fixes them all.

* Add a CHANGELOG.md entry

* Fix a race in the remote-settings integration test

On slow machines the test would fail because it would
take more time to update from the remote settings.