Store "pending consent" and gate joining studies on it
We don't mark studies as active unless user has consented to them.
This is to prevent side-loaded studies to suddenly start running
without user approval. How does this work?

1. Whenever a study is consented and installation is triggered from
   the core add-on UI, we record a "pending consent" (i.e. user
   consented but the study is not yet installed).
2. When a study is installed we detect it and check if we
   had a prior "pending consent" for it. If so, join the study.
3. Joined installed studies are marked as such in the UI. If a study
   is installed, but not joined, they get uninstalled at the earliest
   opportunity by the core add-on.

Note that pending consent is cleared whenever the core add-on is initialized
again, so that pending consent is not retained across browser restarts.