Look for the new label for confirming add-on installation
In bug https://bugzilla.mozilla.org/show_bug.cgi?id=1697622 the
label of the confirmation button was changed, making our tests look
for the wrong element. This updates the tests to look for the newest
label. A more robust approach that looks at the id can be implemented
in future, separate PRs.