Flip upload to true after enrollment
The Glean.initialize() method is called when
the DataCollection.js class is instantiated, early
at the addon startup. On the first run though, due
to the fact that user has to go through the onboarding,
Glean.initialize() is called with `upload=false`.

Without this commit, on the first run and after a successful
onboarding, no ping can be submitted, because collection
is never turned on.